Change Scl_Pair_t_ member types to long to allow for large liberty file loading.

+++ src/map/scl/sclLiberty.c
@@ -49,8 +49,8 @@ typedef enum {
 typedef struct Scl_Pair_t_ Scl_Pair_t;
 struct Scl_Pair_t_
 {
-    int             Beg;          // item beginning
-    int             End;          // item end
+    long            Beg;          // item beginning
+    long            End;          // item end
 };
 
 typedef struct Scl_Item_t_ Scl_Item_t;
